+++ R/energy_balance.R
@@ -4,7 +4,335 @@
 #
 
 
-#' Confirm that an SUT-style data frame conserves energy.
+
+#' Calculate inter- and intra-industry balances for PSUT (**RUVY**) matrices2
+#'
+#' Balances are an important aspect of analyzing mass and energy conversion chains
+#' in the PSUT framework with the **RUVY** matrices.
+#' Often, balances are calculated on energy or exergy flows.
+#' Balances can be calculated for mass flows or monetary flows, too.
+#'
+#' ## Inter-industry balances
+#'
+#' In a PSUT description of a mass or energy conversion chain,
+#' all of every product leaving one industry must arrive at another industry.
+#' `calc_inter_industry_balances()`
+#' calculates these between-industry balances via
+#' ((**R** + **V**)^T^ - (**U** + **Y**))**i**.
+#' Inter-industry balances are calculated for products (not industries).
+#'
+#' ## Intra-industry balances
+#'
+#' `calc_intra_industry_balances()` calculates across-industry
+#' balances via
+#' (**U**^T^ - **V**)**i**.
+#' Inter-industry balances are calculated for industries (not products),
+#' and the result is a column vector of industry balances.
+#'
+#' ## Outputs
+#'
+#' The argument `balance_colname` specifies the name of the
+#' result.
+#' By default, the output of `calc_inter_industry_balances()` is
+#' [Recca::balance_cols]`$inter_industry_balance_colname` or
+#' "`r Recca::balance_cols$inter_industry_balance_colname`".
+#' By default, the output of `calc_intra_industry_balances()` is
+#' [Recca::balance_cols]`$intra_industry_balance_colname` or
+#' "`r Recca::balance_cols$intra_industry_balance_colname`".
+#'
+#' For inter-industry (between-industry) balances,
+#' the result is a column vector of product balances.
+#' The result should __always__ be the **0** vector.
+#'
+#' In a PSUT description of a mass or energy conversion chain,
+#' the meaning of intra-industry (across-industry) balances
+#' depends on the construction of the matrices.
+#' When all losses are accounted in the matrices themselves,
+#' the calculation of intra-industry balances for
+#' mass and energy conversion chains should give the **0** vector
+#' with industries in rows,
+#' because mass and energy are conserved.
+#' When losses are __not__ accounted in the matrices,
+#' the calculation of intra-industry balances gives losses.
+#'
+#' For exergy conversion chains
+#' (mass, energy, or both),
+#' when losses are accounted in the matrices,
+#' the intra-industry balance gives exergy destruction.
+#' When losses are __not__ accounted in the matrices,
+#' the intra-industry balance gives the sum of
+#' exergy destruction and exergy losses.
+#'
+#' @param .sutmats A named list of matrices or
+#'                 an SUT-style, wide-by-matrices data frame
+#'                 with columns of matrices,
+#'                 including `R`, `U`, `V`, and `Y`.
+#' @param R Resources (**R**) matrix or name of the column in `.sutmats`
+#'          that contains same.
+#'          Default is "R".
+#' @param U Use (**U**) matrix or name of the column in `.sutmats`
+#'          that contains same.
+#'          Default is "U".
+#' @param V Make (**V**) matrix or name of the column in `.sutmats`
+#'          that contains same.
+#'          Default is "V".
+#' @param Y Final demand (**Y**) matrix or name of the column in `.sutmats`
+#'          that contains same.
+#'          Default is "Y".
+#' @param balance_colname The name of the column containing energy balance vectors.
+#'                        Defaults are
+#'                        [Recca::balance_cols]`$inter_industry_balance_colname` or
+#'                        "`r Recca::balance_cols$inter_industry_balance_colname`"
+#'                        for
+#'                        `calc_inter_industry_balances()` and
+#'                        [Recca::balance_cols]`$intra_industry_balance_colname` or
+#'                        "`r Recca::balance_cols$intra_industry_balance_colname`"
+#'                        for
+#'                        `calc_intra_industry_balances()`.
+#'
+#' @returns A list or data frame containing balances.
+#'
+#' @seealso [verify_inter_industry_balance()] and [verify_intra_industry_balance()]
+#'
+#' @examples
+#' # Inter-industry balances
+#' inter <- UKEnergy2000mats |>
+#'   dplyr::filter(LastStage %in% c("Final", "Useful")) |>
+#'   tidyr::pivot_wider(names_from = matrix.name, values_from = matrix) |>
+#'   calc_inter_industry_balance()
+#' inter[[Recca::balance_cols$inter_industry_balance_colname]][[1]]
+#' inter[[Recca::balance_cols$inter_industry_balance_colname]][[2]]
+#' # Intra-industry balances
+#' intra <- UKEnergy2000mats |>
+#'   dplyr::filter(LastStage %in% c("Final", "Useful")) |>
+#'   tidyr::pivot_wider(names_from = matrix.name, values_from = matrix) |>
+#'   calc_intra_industry_balance()
+#' intra[[Recca::balance_cols$intra_industry_balance_colname]][[1]]
+#' intra[[Recca::balance_cols$intra_industry_balance_colname]][[2]]
+#' @name balances
+NULL
+
+
+
+#' Verify inter- and intra-industry balances for PSUT (**RUVY**) matrices
+#'
+#' Balances can be verified in two ways,
+#' inter-industry and intra-industry.
+#' Inter-industry balances are between industries and performed by product.
+#' Intra-industry balances are across industries and performed per industry.
+#'
+#' ## Inter-industry balances (by product)
+#'
+#' In a PSUT description of an economy, all of every product
+#' leaving one industry must arrive at another industry.
+#' Inter-industry (between-industry) balances are verified by product
+#' (within `tol`) for every row in `.sutmats`.
+#' Inter-industry balances should be calculated via
+#' `calc_inter_industry_balances()` before calling
+#' `verify_inter_industry_balances()`.
+#' See examples.
+#'
+#' ## Intra-industry balances (by industry)
+#'
+#' Intra-industry (across-industry) balances are
+#' verified by industry (within `tol`) for every row in `.sutmats`.
+#' Intra-industry balances should be calculated via
+#' `calc_intra_industry_balances()` before calling
+#' `verify_intra_industry_balances()`.
+#' See examples.
+#'
+#' The calculation of inter-industry balances
+#' will give non-zero vectors for
+#' quantities that are not conserved (such as exergy)
+#' and for conversion chains that do not include wastes or losses.
+#'
+#' ## Outputs
+#'
+#' If every conversion chain in `.sutmats` is balanced,
+#' `.sutmats` is returned with an additional column
+#' (called `balanced_colname`), and
+#' execution returns to the caller.
+#' If balance is not observed for one or more of the rows
+#' in `.sutmats`,
+#' a warning is emitted, and
+#' the additional column (`balanced_colname`)
+#' indicates where the problem occurred, with `FALSE`
+#' showing where products are not balanced
+#' to within `tol`.
+#'
+#' Typically, one would call `calc_int*_industry_balance()`
+#' before calling `verify_int*_industry_balance()`.
+#' See the examples.
+#'
+#' @param .sutmats An SUT-style data frame with a column
+#'                 named `balances`.
+#' @param balances The name of a column that contains balance vectors.
+#'                 For `verify_inter_industry_balances()`,
+#'                 the default is [Recca::balance_cols]`$inter_industry_balance_colname`
+#'                 or "`r Recca::balance_cols$inter_industry_balance_colname`".
+#'                 For `verify_intra_industry_balances()`,
+#'                 the default is [Recca::balance_cols]`$intra_industry_balance_colname`
+#'                 or "`r Recca::balance_cols$intra_industry_balance_colname`".
+#' @param tol The maximum amount by which products can be out of balance
+#'            and still be considered balanced.
+#'            Default is `1e-6`.
+#' @param balanced_colname The name for booleans telling if balance is present.
+#'                         For `verify_inter_industry_balances()`,
+#'                         the default is [Recca::balance_cols]`$inter_industry_balanced_colname`
+#'                         or "`r Recca::balance_cols$inter_industry_balanced_colname`".
+#'                         For `verify_intra_industry_balances()`,
+#'                         the default is [Recca::balance_cols]`$intra_industry_balanced_colname`
+#'                         or "`r Recca::balance_cols$intra_industry_balanced_colname`".
+#'
+#' @return A list or data frame with an additional value or column
+#'         saying whether `.sutmats` are in balance.
+#'
+#' @seealso [calc_inter_industry_balance()], [calc_intra_industry_balance()]
+#'
+#' @examples
+#' library(dplyr)
+#' library(tidyr)
+#' result_inter <- UKEnergy2000mats |>
+#'   dplyr::filter(LastStage %in% c("Final", "Useful")) |>
+#'   tidyr::pivot_wider(names_from = matrix.name,
+#'                      values_from = matrix) |>
+#'   calc_inter_industry_balance() |>
+#'   verify_inter_industry_balance(tol = 1e-4)
+#' result_inter
+#' result_inter[[Recca::balance_cols$inter_industry_balanced_colname]]
+#' @name verify-balances
+NULL
+
+
+
+#' @export
+#' @rdname balances
+calc_inter_industry_balance <- function(.sutmats = NULL,
+                                        # Input names
+                                        R = "R",
+                                        U = "U",
+                                        V = "V",
+                                        Y = "Y",
+                                        # Output name
+                                        balance_colname = Recca::balance_cols$inter_industry_balance_colname) {
+  calc_bal_func <- function(R_mat = NULL, U_mat, V_mat, Y_mat) {
+    V_sums <- matsbyname::colsums_byname(V_mat)
+    # Allow for missing R matrix
+    if (is.null(R_mat)) {
+      RV_sums <- V_sums
+    } else {
+      R_sums <- matsbyname::colsums_byname(R_mat)
+      RV_sums <- matsbyname::sum_byname(R_sums, V_sums)
+    }
+    RV_sums <- matsbyname::transpose_byname(RV_sums)
+    U_sums <- matsbyname::rowsums_byname(U_mat)
+    Y_sums <- matsbyname::rowsums_byname(Y_mat)
+    UY_sums <- matsbyname::sum_byname(U_sums, Y_sums)
+    # (R + V)^T - (U + Y)
+    bal <- matsbyname::difference_byname(RV_sums, UY_sums)
+    list(bal) |>
+      magrittr::set_names(balance_colname)
+  }
+  matsindf::matsindf_apply(.sutmats, FUN = calc_bal_func, R_mat = R, U_mat = U, V_mat = V, Y_mat = Y)
+}
+
+
+
+#' @export
+#' @rdname verify-balances
+verify_inter_industry_balance <- function(.sutmats = NULL,
+                                          # Input names
+                                          balances = Recca::balance_cols$inter_industry_balance_colname,
+                                          # Tolerance
+                                          tol = 1e-6,
+                                          # Output name
+                                          balanced_colname = Recca::balance_cols$inter_industry_balanced_colname) {
+  verify_func_inter <- function(bal_vector) {
+    OK <- bal_vector |>
+      matsbyname::iszero_byname(tol) |>
+      as.logical()
+    if (!OK) {
+      return(list(FALSE) |>
+               magrittr::set_names(balanced_colname))
+    }
+    list(TRUE) |>
+      magrittr::set_names(balanced_colname)
+  }
+
+  out <- matsindf::matsindf_apply(.sutmats, FUN = verify_func_inter, bal_vector = balances)
+  if (!all(out[[balanced_colname]] %>% as.logical())) {
+    warning(paste0("Products are not conserved in verify_inter_industry_balance(). See column ", balanced_colname, "."))
+  }
+  return(out)
+}
+
+
+
+#' @export
+#' @rdname balances
+calc_intra_industry_balance <- function(.sutmats = NULL,
+                                        # Input names
+                                        U = "U",
+                                        V = "V",
+                                        # Output name
+                                        balance_colname = Recca::balance_cols$intra_industry_balance_colname) {
+  calc_bal_func <- function(U_mat, V_mat) {
+    bal <- matsbyname::difference_byname(matsbyname::transpose_byname(U_mat),
+                                         V_mat) |>
+      matsbyname::rowsums_byname()
+    list(bal) |>
+      magrittr::set_names(balance_colname)
+  }
+  matsindf::matsindf_apply(.sutmats, FUN = calc_bal_func, U_mat = U, V_mat = V)
+}
+
+
+
+#' @export
+#' @rdname verify-balances
+verify_intra_industry_balance <- function(.sutmats = NULL,
+                                          # Input names
+                                          balances = Recca::balance_cols$intra_industry_balance_colname,
+                                          # Tolerance
+                                          tol = 1e-6,
+                                          # Output name
+                                          balanced_colname = Recca::balance_cols$intra_industry_balanced_colname) {
+  verify_func_intra <- function(bal_vector) {
+    OK <- bal_vector |>
+      matsbyname::iszero_byname(tol) |>
+      as.logical()
+    if (!OK) {
+      return(list(FALSE) |>
+               magrittr::set_names(balanced_colname))
+    }
+    list(TRUE) |>
+      magrittr::set_names(balanced_colname)
+  }
+
+  out <- matsindf::matsindf_apply(.sutmats, FUN = verify_func_intra, bal_vector = balances)
+  if (!all(out[[balanced_colname]] |> as.logical())) {
+    warning(paste0("Industries are not balanced in verify_intra_industry_balance(). See column ", balanced_colname, "."))
+  }
+  return(out)
+}
+
+
+
+
+
+
+
+
+
+
+#' Verifies SUT inter-industry energy balances
+#'
+#' `r lifecycle::badge('deprecated')`
+#'
+#' [verify_SUT_energy_balance()] is deprecated in favor of the combination of
+#' [calc_inter_industry_balance()] and
+#' [verify_inter_industry_balance()].
+#' Pipe the first into the second.
 #'
 #' Energy balances are confirmed by Product (within \code{tol}) for every row in \code{.sutmats}.
 #'
@@ -17,26 +345,37 @@
 #' showing where energy is not in balance.
 #'
 #' @param .sutmats an SUT-style data frame with columns of matrices,
-#'        including `U`, `V`, and `Y` columns.
+#'                 including `U`, `V`, and `Y` columns.
 #' @param R resources (`R`) matrix or name of the column in `.sutmats` that contains same. Default is "R".
 #' @param U use (`U`) matrix or name of the column in `.sutmats` that contains same. Default is "U".
 #' @param V make (`V`) matrix or name of the column in `.sutmats`that contains same. Default is "V".
 #' @param Y final demand (`Y`) matrix or name of the column in `.sutmats` that contains same. Default is "Y".
-#' @param SUT_energy_balance the name for booleans telling if energy is in balance. Default is ".SUT_energy_balance".
 #' @param tol the maximum amount by which Supply and Consumption can be out of balance.
-#'        Default is `1e-6`.
+#'            Default is `1e-6`.
+#' @param SUT_energy_balance The name of the column in output containing logical
+#'                           values representing the balance status of that row.
 #'
 #' @return a list or data frame saying whether `.sutmats` are in balance.
 #'
 #' @export
 #'
+#' @keywords internal
+#'
 #' @examples
 #' library(dplyr)
 #' library(tidyr)
-#' verify_SUT_energy_balance(UKEnergy2000mats %>%
-#'                             dplyr::filter(LastStage %in% c("Final", "Useful")) %>%
-#'                             tidyr::spread(key = matrix.name, value = matrix),
-#'                           tol = 1e-4)
+#' # This function is deprecated.
+#' # Instead of this:
+#' UKEnergy2000mats |>
+#'   dplyr::filter(LastStage %in% c("Final", "Useful")) |>
+#'   tidyr::spread(key = matrix.name, value = matrix) |>
+#'   verify_SUT_energy_balance(tol = 1e-4)
+#' # Do this:
+#' UKEnergy2000mats |>
+#'   dplyr::filter(LastStage %in% c("Final", "Useful")) |>
+#'   tidyr::spread(key = matrix.name, value = matrix) |>
+#'   calc_inter_industry_balance() |>
+#'   verify_inter_industry_balance(tol = 1e-4)
 verify_SUT_energy_balance <- function(.sutmats = NULL,
                                       # Input names
                                       R = "R", U = "U", V = "V", Y = "Y",
@@ -44,39 +383,43 @@ verify_SUT_energy_balance <- function(.sutmats = NULL,
                                       tol = 1e-6,
                                       # Output name
                                       SUT_energy_balance = ".SUT_energy_balance"){
-  verify_func <- function(R_mat = NULL, U_mat, V_mat, Y_mat){
-    if (is.null(R_mat)) {
-      # No R matrix, just use the V matrix, assuming that resouces are included there.
-      R_plus_V_mat <- V_mat
-    } else {
-      # An R matrix is present. Sum R and V before proceeding.
-      R_plus_V_mat <- matsbyname::sum_byname(R_mat, V_mat)
-    }
-    RV_sums <- matsbyname::transpose_byname(R_plus_V_mat) %>% matsbyname::rowsums_byname()
-    U_sums <- matsbyname::rowsums_byname(U_mat)
-    Y_sums <- matsbyname::rowsums_byname(Y_mat)
-    # (R + V) - U - Y
-    err <- matsbyname::difference_byname(RV_sums, U_sums) %>% matsbyname::difference_byname(Y_sums)
-    OK <- err %>%
-      matsbyname::iszero_byname(tol) %>%
-      as.logical()
-    if (!OK) {
-      return(list(FALSE) %>% magrittr::set_names(SUT_energy_balance))
-    }
-    list(TRUE) %>% magrittr::set_names(SUT_energy_balance)
-  }
+  lifecycle::deprecate_soft(when = "0.1.65",
+                            what = "verify_SUT_energy_balance()",
+                            with = "verify_inter_industry_balance()")
 
-  Out <- matsindf::matsindf_apply(.sutmats, FUN = verify_func, R_mat = R, U_mat = U, V_mat = V, Y_mat = Y)
-  if (!all(Out[[SUT_energy_balance]] %>% as.logical())) {
-    warning(paste("Energy not conserved in verify_SUT_energy_balance. See", SUT_energy_balance))
+  out <- .sutmats |>
+    calc_inter_industry_balance(R = R, U = U, V = V, Y = Y,
+                                balance_colname = Recca::balance_cols$inter_industry_balance_colname) |>
+    verify_inter_industry_balance(tol = tol,
+                                  balances = Recca::balance_cols$inter_industry_balance_colname,
+                                  balanced_colname = SUT_energy_balance)
+  # Remove the balances item to mimic the behaviour of the deprecated function.
+  if (is.data.frame(out)) {
+    out <- out |>
+      dplyr::mutate(
+        "{Recca::balance_cols$inter_industry_balance_colname}" := NULL
+      )
+  } else {
+    # Assume out is a list
+    out[Recca::balance_cols$inter_industry_balance_colname] <- NULL
   }
-  return(Out)
+  return(out)
 }
+
+
+
+
+
+
+
 
 
 #' Confirm that an SUT-style data frame conserves energy.
 #'
-#' If energy is in balance for every row, `.sutmats` is returned with two additional columns, and
+#' `r lifecycle::badge('deprecated')`
+#'
+#' If energy is in balance for every row,
+#' `.sutmats` is returned with two additional columns, and
 #' execution returns to the caller.
 #' If energy balance is not observed for one or more rows,
 #' a warning is emitted, and
@@ -107,6 +450,8 @@ verify_SUT_energy_balance <- function(.sutmats = NULL,
 #'
 #' @export
 #'
+#' @keywords internal
+#'
 #' @examples
 #' library(tidyr)
 #' verify_SUT_energy_balance_with_units(UKEnergy2000mats %>%
@@ -131,7 +476,12 @@ verify_SUT_energy_balance_with_units <- function(.sutmats = NULL,
                                                  SUT_prod_energy_balanced = ".SUT_prod_energy_balanced",
                                                  SUT_ind_energy_balanced = ".SUT_ind_energy_balanced",
                                                  ebal_error = "ebal_error",
-                                                 product = "Product"){
+                                                 product = "Product") {
+
+  lifecycle::deprecate_warn(when = "0.1.65",
+                            what = "verify_SUT_energy_balance_with_units()",
+                            details = "verify_SUT_energy_balance_with_units() seemed like a good idea at the time, but we never use it.")
+
   verify_func <- function(R = NULL, U, V, Y, S_units){
     y <- matsbyname::rowsums_byname(Y)
     if (is.null(R)) {
@@ -373,4 +723,90 @@ verify_IEATable_energy_balance <- function(.ieatidydata,
                           msg = paste("Energy not balanced in verify_IEATable_energy_balance."))
 
   return(EnergyCheck)
+}
+
+
+
+#' Endogenize losses into PSUT matrices
+#'
+#' When a conversion chain does _not_ include losses in the
+#' **RUVY** matrices of the PSUT framework,
+#' it may be helpful to endogenize the losses.
+#' This function performs the endogenization.
+#'
+#' Intra-industry balances need to be calculated
+#' (most easily via [calc_intra_industry_balance()])
+#' prior to calling this function.
+#' See the example.
+#'
+#' @param .sutmats A `matsindf` data frame, wide by matrices.
+#' @param V Make (**V**) matrix or name of the column in `.sutmats`
+#'          that contains same.
+#'          Default is "V".
+#' @param Y Final demand (**Y**) matrix or name
+#'          of the column in `.sutmats` that contains same.
+#'          Default is "Y".
+#' @param loss_product The string name of the loss product.
+#'                     Default is [Recca::balance_cols]`$waste_heat` or
+#'                     "`r Recca::balance_cols$waste_heat`".
+#' @param loss_sector The string name of the sector
+#'                    that will absorb losses.
+#'                    Default is [Recca::balance_cols]`$losses_sector` or
+#'                    "`r Recca::balance_cols$losses_sector`".
+#' @param balance_colname The name of the column containing
+#'                        energy balance vectors.
+#'                        Default is
+#'                        [Recca::balance_cols]`$intra_industry_balance_colname` or
+#'                        "`r Recca::balance_cols$intra_industry_balance_colname`".
+#' @param V_prime The name of the **V** matrix with endogenized losses.
+#' @param Y_prime The name of the **Y** matrix with endogenized losses.
+#'
+#' @returns A version of `.sutmats` with losses endogenized.
+#'
+#' @export
+#'
+#' @examples
+#' UKEnergy2000mats |>
+#'   tidyr::pivot_wider(names_from = matrix.name,
+#'                      values_from = matrix) |>
+#'   dplyr::filter(.data[[IEATools::iea_cols$last_stage]] %in%
+#'                   c(IEATools::last_stages$final, IEATools::last_stages$useful)) |>
+#'   calc_intra_industry_balance() |>
+#'   endogenize_losses() |>
+#'   # Use the endogenized matrices
+#'   dplyr::mutate(
+#'     V = V_prime,
+#'     Y = Y_prime,
+#'     V_prime = NULL,
+#'     Y_prime = NULL,
+#'     "{Recca::balance_cols$intra_industry_balance_colname}" := NULL
+#'   )
+endogenize_losses <- function(.sutmats = NULL,
+                              loss_product = Recca::balance_cols$waste_heat,
+                              loss_sector = Recca::balance_cols$losses_sector,
+                              V = "V",
+                              Y = "Y",
+                              balance_colname = Recca::balance_cols$intra_industry_balance_colname,
+                              # Output columns
+                              V_prime = "V_prime",
+                              Y_prime = "Y_prime") {
+
+  endogenize_func <- function(V_mat, Y_mat, balance_vec) {
+    add_to_V <- balance_vec |>
+      matsbyname::setcolnames_byname(loss_product)
+    V_prime_mat <- matsbyname::sum_byname(V_mat, add_to_V)
+    add_to_Y <- balance_vec |>
+      matsbyname::transpose_byname() |>
+      matsbyname::sumall_byname() |>
+      matsbyname::setrownames_byname(loss_product) |>
+      matsbyname::setcolnames_byname(loss_sector)
+    Y_prime_mat <- matsbyname::sum_byname(Y_mat, add_to_Y)
+    list(V_prime_mat, Y_prime_mat) |>
+      magrittr::set_names(c(V_prime, Y_prime))
+  }
+  matsindf::matsindf_apply(.sutmats,
+                           FUN = endogenize_func,
+                           V_mat = V,
+                           Y_mat = Y,
+                           balance_vec = balance_colname)
 }